+++ demisto_sdk/commands/common/errors.py
@@ -116,7 +116,7 @@ ERROR_CODE = {
     "invalid_deprecated_playbook": "PB104",
     "playbook_cant_have_deletecontext_all": "PB105",
     "using_instance_in_playbook": "PB106",
-    "invalid_script_id": "PB107",
+    "playbook_non_existing_scripts": "PB107",
     "description_missing_in_beta_integration": "DS100",
     "no_beta_disclaimer_in_description": "DS101",
     "no_beta_disclaimer_in_yml": "DS102",
@@ -781,12 +781,8 @@ class Errors:
 
     @staticmethod
     @error_code_decorator
-    def invalid_script_id(script_entry_to_check, pb_task):
-        return f"in task {pb_task} the script {script_entry_to_check} was not found in the id_set.json file. " \
-               f"Please make sure:\n" \
-               f"1 - The right script id is set and the spelling is correct.\n" \
-               f"2 - The id_set.json file is up to date. Delete the file by running: rm -rf Tests/id_set.json and" \
-               f" rerun the command."
+    def playbook_non_existing_scripts(implemented_scripts_in_playbook):
+        return f"The following scripts: {implemented_scripts_in_playbook} do not exist."
 
     @staticmethod
     @error_code_decorator