+++ demisto_sdk/commands/common/tests/playbook_test.py
@@ -223,55 +223,6 @@ class TestPlaybookValidator:
         (NEXT_TASKS_VALID_EXIST_2, True),
     ]
 
-    PLAYBOOK_JSON_VALID_SCRIPT_ID = {
-        "tasks": {"0": {"task": {"script": "scriptId1"}},
-                  "1": {"task": {"script": "scriptId2"}}}}
-    ID_SET_VALID_SCRIPT_ID = {"scripts": [{"scriptId1": {"name": "name"}}, {"scriptId2": {"name": "name"}}]}
-
-    PLAYBOOK_JSON_INVALID_SCRIPT_ID = {
-        "tasks": {"0": {"task": {"script": "scriptId1"}},
-                  "1": {"task": {"script": "scriptId2"}}}}
-    ID_SET_INVALID_SCRIPT_ID = {"scripts": [{"scriptId2": {"name": "name"}}]}
-
-    PLAYBOOK_JSON_VALID_SCRIPT_NAME = {
-        "tasks": {"0": {"task": {"scriptName": "scriptName1"}},
-                  "1": {"task": {"scriptName": "scriptName2"}}}}
-    ID_SET_VALID_SCRIPT_NAME = {
-        "scripts": [{"scriptId1": {"name": "scriptName1"}}, {"scriptId2": {"name": "scriptName2"}}]}
-
-    PLAYBOOK_JSON_INVALID_SCRIPT_NAME = {
-        "tasks": {"0": {"task": {"scriptName": "scriptName1"}},
-                  "1": {"task": {"scriptName": "scriptName2"}}}}
-    ID_SET_INVALID_SCRIPT_NAME = {
-        "scripts": [{"scriptId1": {"name": "scriptName3"}}, {"scriptId2": {"name": "scriptName1"}}]}
-
-    IS_SCRIPT_ID_VALID = [
-        (PLAYBOOK_JSON_VALID_SCRIPT_ID, ID_SET_VALID_SCRIPT_ID, True),
-        (PLAYBOOK_JSON_INVALID_SCRIPT_ID, ID_SET_INVALID_SCRIPT_ID, False),
-        (PLAYBOOK_JSON_VALID_SCRIPT_NAME, ID_SET_VALID_SCRIPT_NAME, True),
-        (PLAYBOOK_JSON_INVALID_SCRIPT_NAME, ID_SET_INVALID_SCRIPT_NAME, False),
-    ]
-
-    @pytest.mark.parametrize("playbook_json, id_set_json, expected_result", IS_SCRIPT_ID_VALID)
-    def test_playbook_script_id(self, mocker, playbook, repo, playbook_json, id_set_json, expected_result):
-        """
-
-        Given
-        - A playbook with scrips ids or script names
-        - An id_set file.
-
-        When
-        - validating playbook
-
-        Then
-        - In case script id or script name don't exist in id_set , prints a warning.
-        """
-        playbook.yml.write_dict(playbook_json)
-        repo.id_set.write_json(id_set_json)
-        structure = mock_structure("", playbook_json)
-        validator = PlaybookValidator(structure)
-        assert validator.is_script_id_valid(id_set_json) == expected_result
-
     @pytest.mark.parametrize("playbook_json, expected_result", IS_NO_ROLENAME_INPUTS)
     def test_is_added_required_fields(self, playbook_json, expected_result):
         structure = mock_structure("", playbook_json)
@@ -292,7 +243,7 @@ class TestPlaybookValidator:
 
     @pytest.mark.parametrize("playbook_path, expected_result", [(INVALID_TEST_PLAYBOOK_UNHANDLED_CONDITION, True),
                                                                 (INVALID_PLAYBOOK_UNHANDLED_CONDITION, False)])
-    def test_skipping_test_playbooks(self, mocker, playbook_path, expected_result):
+    def test_skipping_test_playbooks(self, playbook_path, expected_result):
         """
             Given
             - A playbook
@@ -306,7 +257,6 @@ class TestPlaybookValidator:
         """
         structure = StructureValidator(file_path=playbook_path)
         validator = PlaybookValidator(structure)
-        mocker.patch.object(validator, 'is_script_id_valid', return_value=True)
         assert validator.is_valid_playbook() is expected_result
 
     @pytest.mark.parametrize("playbook_json, expected_result", IS_DELETECONTEXT)